have add league also add hero entries for league

+++ app/models/LeagueUser.scala
@@ -7,7 +7,7 @@ import java.sql.Timestamp
 class LeagueUser(
                   val leagueId: Int,
                   val userId: Int,
-                  var money: Double,
+                  var money: BigDecimal,
                   var entered: Timestamp,
                   var remainingTransfers: Int,
                   var changeTstamp: Option[Timestamp]
@@ -20,7 +20,7 @@ class LeagueUserStats(
                        val statFieldId: Long,
                        val leagueUserId: Long,
                        val day: Int,
-                       var value: Double,
+                       var value: BigDecimal,
                        var oldRank: Int = 0,
                      ) extends KeyedEntity[Long] {
   val id: Long = 0
